style(foldersystemconfig): move config set-up method
move config setupmethod from FolderSystem  to FolderSystemCOnfig class
