+++ src/libs/Config/FolderSystemConfig.ts
@@ -1,14 +1,18 @@
+import { existsSync, mkdirSync } from 'fs';
+import path from 'path';
 import { DefaultConfig } from './DefaultConfig';
 
-export type FolderSystemConfigProperties = {
-  location: string;
-  rootBucket: string;
-  idmlFileBucket: string;
-  zipFileBucket: string;
-  unzipFileBucket: string;
-};
 export class FolderSystemConfig extends DefaultConfig {
-  getConfig(): FolderSystemConfigProperties {
-    return this.getConfigProperties() as FolderSystemConfigProperties;
+  async configSetUp(): Promise<void> {
+    const { location, rootBucket, idmlFileBucket, zipFileBucket, unzipFileBucket } = this.getConfigProperties();
+    const rootPath = path.join(location, rootBucket);
+    const folders = [idmlFileBucket, zipFileBucket, unzipFileBucket];
+
+    folders.forEach((folder) => {
+      const folderPath = path.join(rootPath, folder);
+      if (!existsSync(folderPath)) {
+        mkdirSync(folderPath, { recursive: true });
+      }
+    });
   }
 }
