+++ src/libs/FolderSystem/FolderSystem.ts
@@ -1,4 +1,3 @@
-import { existsSync, mkdirSync } from 'fs';
 import path from 'path';
 import { FolderSystemConfig } from '../Config/FolderSystemConfig';
 import { FilePaths, FolderSystemFilePaths, IFolderSystem } from './IFolderSystem';
@@ -14,26 +13,16 @@ export class FolderSystem implements IFolderSystem {
 
   private unzipFileBucket: string;
 
-  constructor(private filename: string, private folderSystemConfig: FolderSystemConfig = new FolderSystemConfig()) {
+  private folderSystemConfig: FolderSystemConfig = new FolderSystemConfig();
+
+  constructor(private filename: string) {
     const { location, rootBucket, idmlFileBucket, zipFileBucket, unzipFileBucket } =
-      this.folderSystemConfig.getConfig();
+      this.folderSystemConfig.getConfigProperties();
     this.location = location;
     this.rootBucket = rootBucket;
     this.idmlFileBucket = idmlFileBucket;
     this.zipFileBucket = zipFileBucket;
     this.unzipFileBucket = unzipFileBucket;
-  }
-
-  async configSetUp(): Promise<void> {
-    const rootPath = path.join(this.location, this.rootBucket);
-    const folders = [this.idmlFileBucket, this.zipFileBucket, this.unzipFileBucket];
-
-    folders.forEach((folder) => {
-      const folderPath = path.join(rootPath, folder);
-      if (!existsSync(folderPath)) {
-        mkdirSync(folderPath, { recursive: true });
-      }
-    });
   }
 
   getFilePaths(): FolderSystemFilePaths {
